remove extra deps, fix bug with long values

+++ src/main/java/ru/kduskov/filedatafilter/enums/ContentType.java
@@ -2,6 +2,6 @@ package ru.kduskov.filedatafilter.enums;
 
 public enum ContentType {
     STRING,
-    INT,
+    LONG,
     FLOAT
 }
